+++ assets/src/js/app.js
@@ -2,6 +2,8 @@ import navigation from './navigation';
 
 import scroll from './scroll';
 
+import scrollAnimation from './scrollAnimation';
+
 import carousel from './carousel';
 
 import validation from './validation';
@@ -11,6 +13,7 @@ import validation from './validation';
     const scripts = [
         navigation,
         scroll,
+        scrollAnimation,
         carousel,
         validation
     ];